+++ tests/test_codesyncd/populate_buffer.test.js
@@ -9,9 +9,11 @@ import {pathUtils} from "../../src/utils/path_utils";
 import {readYML} from "../../src/utils/common";
 import {populateBuffer} from "../../src/codesyncd/populate_buffer";
 import {createSystemDirectories} from "../../src/utils/setup_utils";
-import {DEFAULT_BRANCH, DIFF_SOURCE} from "../../src/constants";
+import {DEFAULT_BRANCH} from "../../src/constants";
 import {
     assertChangeEvent,
+    assertNewFileEvent,
+    assertFileDeleteEvent,
     assertRenameEvent,
     DUMMY_FILE_CONTENT,
     getConfigFilePath,
@@ -90,27 +92,10 @@ describe("populateBuffer", () => {
         fs.writeFileSync(userFilePath, yaml.safeDump(userData));
     };
 
-    const assertNewFileEvent = (newRelPath, diffsCount = 1) => {
-        const originalsFilePath = path.join(originalsRepoBranchPath, newRelPath);
-        const shadowFilePath = path.join(shadowRepoBranchPath, newRelPath);
-        // Verify file has been created in the .shadow repo and .originals repos
-        expect(fs.existsSync(shadowFilePath)).toBe(true);
-        expect(fs.existsSync(originalsFilePath)).toBe(true);
-        // Verify correct diff file has been generated
-        let diffFiles = fs.readdirSync(diffsRepo);
-        expect(diffFiles).toHaveLength(diffsCount);
-        const diffFilePath = path.join(diffsRepo, diffFiles[diffsCount-1]);
-        const diffData = readYML(diffFilePath);
-        expect(diffData.source).toEqual(DIFF_SOURCE);
-        expect(diffData.is_new_file).toBe(true);
-        expect(diffData.is_rename).toBeFalsy();
-        expect(diffData.is_deleted).toBeFalsy();
-        expect(diffData.repo_path).toEqual(repoPath);
-        expect(diffData.branch).toEqual(DEFAULT_BRANCH);
-        expect(diffData.file_relative_path).toEqual(newRelPath);
-        expect(diffData.diff).toEqual("");
-        return true;
-    };
+    afterEach(() => {
+        fs.rmSync(repoPath, {recursive: true, force: true});
+        fs.rmSync(baseRepoPath, {recursive: true, force: true});
+    });
 
     test("No repo synced", async () => {
         await populateBuffer();
@@ -149,7 +134,7 @@ describe("populateBuffer", () => {
         addRepo();
         fs.writeFileSync(newFilePath, DUMMY_FILE_CONTENT);
         await populateBuffer();
-        expect(assertNewFileEvent("new.js")).toBe(true);
+        expect(assertNewFileEvent(repoPath, "new.js")).toBe(true);
     });
 
     test("New Binary File", async () => {
@@ -165,7 +150,7 @@ describe("populateBuffer", () => {
         var buf = Buffer.from(data, 'base64');
         fs.writeFileSync(newFilePath, buf);
         await populateBuffer();
-        expect(assertNewFileEvent(newRelPath)).toBe(true);
+        expect(assertNewFileEvent(repoPath, newRelPath)).toBe(true);
     });
 
     test("Rename event", async () => {
@@ -186,29 +171,14 @@ describe("populateBuffer", () => {
         const renamedPath = path.join(repoPath, newRelPath);
         fs.renameSync(filePath, renamedPath);
         await populateBuffer();
-        expect(assertNewFileEvent(newRelPath)).toBe(true);
+        expect(assertNewFileEvent(repoPath, newRelPath)).toBe(true);
     });
 
     test("Delete event", async () => {
         addRepo();
         fs.writeFileSync(shadowFilePath, DUMMY_FILE_CONTENT);
         await populateBuffer();
-        const cacheFilePath = path.join(cacheRepoBranchPath, fileRelPath);
-        // Verify that file is copied to .delete directory
-        expect(fs.existsSync(cacheFilePath)).toBe(true);
-        // Verify correct diff file has been generated
-        let diffFiles = fs.readdirSync(diffsRepo);
-        expect(diffFiles).toHaveLength(1);
-        const diffFilePath = path.join(diffsRepo, diffFiles[0]);
-        const diffData = readYML(diffFilePath);
-        expect(diffData.source).toEqual(DIFF_SOURCE);
-        expect(diffData.is_rename).toBeFalsy();
-        expect(diffData.is_new_file).toBeFalsy();
-        expect(diffData.is_deleted).toBe(true);
-        expect(diffData.repo_path).toEqual(repoPath);
-        expect(diffData.branch).toEqual(DEFAULT_BRANCH);
-        expect(diffData.file_relative_path).toEqual(fileRelPath);
-        expect(diffData.diff).toStrictEqual("");
+        expect(assertFileDeleteEvent(repoPath, fileRelPath)).toBe(true);
     });
 
     test("New File -> Edit -> Rename -> Edit", async () => {
@@ -217,7 +187,7 @@ describe("populateBuffer", () => {
         fs.writeFileSync(filePath, DUMMY_FILE_CONTENT);
         await populateBuffer();
         await waitFor(0.1);
-        expect(assertNewFileEvent(fileRelPath)).toBe(true);
+        expect(assertNewFileEvent(repoPath, fileRelPath)).toBe(true);
         // Edit
         let updatedText = `${DUMMY_FILE_CONTENT} Changed data`;
         fs.writeFileSync(filePath, updatedText);