+++ tests/test_events/handler/handleDelete.test.js
@@ -4,17 +4,18 @@ import vscode from "vscode";
 import untildify from "untildify";
 import getBranchName from "current-git-branch";
 
-import {readYML} from "../../../src/utils/common";
 import {pathUtils} from "../../../src/utils/path_utils";
 import {eventHandler} from "../../../src/events/event_handler";
-import {DEFAULT_BRANCH, DIFF_SOURCE} from "../../../src/constants";
+import {DEFAULT_BRANCH} from "../../../src/constants";
 import {
+    assertFileDeleteEvent,
     Config,
     getConfigFilePath,
     randomBaseRepoPath,
     randomRepoPath,
     waitFor
 } from "../../helpers/helpers";
+import {populateBuffer} from "../../../src/codesyncd/populate_buffer";
 
 describe("handleDeletedEvent",  () => {
     /*
@@ -38,19 +39,19 @@ describe("handleDeletedEvent",  () => {
     const shadowRepoBranchPath = pathUtilsObj.getShadowRepoBranchPath();
     const diffsRepo = pathUtilsObj.getDiffsRepo();
 
+    const fileRelPath = "file_1.js";
     // For file
-    const filePath = path.join(repoPath, "file.js");
+    const filePath = path.join(repoPath, fileRelPath);
     const cacheRepoBranchPath = pathUtilsObj.getDeletedRepoBranchPath();
-    const cacheFilePath = path.join(cacheRepoBranchPath, "file.js");
-    const shadowFilePath = path.join(shadowRepoBranchPath, "file.js");
+    const cacheFilePath = path.join(cacheRepoBranchPath, fileRelPath);
+    const shadowFilePath = path.join(shadowRepoBranchPath, fileRelPath);
 
     // For directory
     const directoryPath = path.join(repoPath, "directory");
-    const directoryFilePath = path.join(directoryPath, "file.js");
-    const relFilePath = path.join("directory", "file.js");
+    const directoryFilePath = path.join(directoryPath, fileRelPath);
+    const relFilePath = path.join("directory", fileRelPath);
     const shadowDirectoryPath = path.join(shadowRepoBranchPath, "directory");
-    const shadowDirectoryFilePath = path.join(shadowDirectoryPath, "file.js");
-    const cacheDirectoryPath = path.join(cacheRepoBranchPath, "directory");
+    const shadowDirectoryFilePath = path.join(shadowDirectoryPath, fileRelPath);
 
     beforeEach(() => {
         jest.clearAllMocks();
@@ -123,21 +124,21 @@ describe("handleDeletedEvent",  () => {
         };
         const handler = new eventHandler();
         handler.handleDeleteEvent(event);
-        // Verify that file is copied to .delete directory
-        expect(fs.existsSync(cacheFilePath)).toBe(true);
-        // Verify correct diff file has been generated
-        let diffFiles = fs.readdirSync(diffsRepo);
-        expect(diffFiles).toHaveLength(1);
-        const diffFilePath = path.join(diffsRepo, diffFiles[0]);
-        const diffData = readYML(diffFilePath);
-        expect(diffData.source).toEqual(DIFF_SOURCE);
-        expect(diffData.is_rename).toBeFalsy();
-        expect(diffData.is_new_file).toBeFalsy();
-        expect(diffData.is_deleted).toBe(true);
-        expect(diffData.repo_path).toEqual(repoPath);
-        expect(diffData.branch).toEqual(DEFAULT_BRANCH);
-        expect(diffData.file_relative_path).toEqual("file.js");
-        expect(diffData.diff).toStrictEqual("");
+        expect(assertFileDeleteEvent(repoPath, fileRelPath)).toBe(true);
+    });
+
+    test("With Daemon: Repo synced, shadow exists",  async () => {
+        const event = {
+            files: [{
+                fsPath: filePath,
+                path: filePath,
+                scheme: "file"
+            }]
+        };
+        const handler = new eventHandler();
+        handler.handleDeleteEvent(event);
+        await populateBuffer();
+        expect(assertFileDeleteEvent(repoPath, fileRelPath)).toBe(true);
     });
 
     test("Repo synced, shadow does NOT exists",  () => {
@@ -187,24 +188,7 @@ describe("handleDeletedEvent",  () => {
         const handler = new eventHandler();
         handler.handleDeleteEvent(event);
         await waitFor(1);
-        // Verify that file is copied to .delete directory
-        expect(fs.existsSync(cacheDirectoryPath)).toBe(true);
-        // Verify correct diff file has been generated
-        let diffFiles = fs.readdirSync(diffsRepo);
-        expect(diffFiles).toHaveLength(1);
-        // Verify correct diff file has been generated
-        expect(diffFiles).toHaveLength(1);
-        const diffFilePath = path.join(diffsRepo, diffFiles[0]);
-        const diffData = readYML(diffFilePath);
-        expect(diffData.source).toEqual(DIFF_SOURCE);
-        expect(diffData.is_deleted).toBe(true);
-        expect(diffData.is_rename).toBeFalsy();
-        expect(diffData.is_new_file).toBeFalsy();
-        expect(diffData.created_at).toBeTruthy();
-        expect(diffData.repo_path).toEqual(repoPath);
-        expect(diffData.branch).toEqual(DEFAULT_BRANCH);
-        expect(diffData.file_relative_path).toEqual(relFilePath);
-        expect(diffData.diff).toEqual("");
+        expect(assertFileDeleteEvent(repoPath, relFilePath, true)).toBe(true);
     });
 });
 
@@ -223,6 +207,7 @@ describe("handleDirectoryDeleteDiffs", () => {
 
     const repoPath = randomRepoPath();
     const baseRepoPath = randomBaseRepoPath();
+    const fileRelPath = "file_1.js";
 
     untildify.mockReturnValue(baseRepoPath);
 
@@ -232,8 +217,8 @@ describe("handleDirectoryDeleteDiffs", () => {
     const cacheRepoBranchPath = pathUtilsObj.getDeletedRepoBranchPath();
 
     const shadowDirectoryPath = path.join(shadowRepoBranchPath, "directory");
-    const shadowFilePath = path.join(shadowDirectoryPath, "file.js");
-    const relFilePath = path.join("directory", "file.js");
+    const shadowFilePath = path.join(shadowDirectoryPath, fileRelPath);
+    const relFilePath = path.join("directory", fileRelPath);
     const cacheFilePath = path.join(cacheRepoBranchPath, relFilePath);
 
     beforeEach(() => {
@@ -254,22 +239,7 @@ describe("handleDirectoryDeleteDiffs", () => {
         const handler = new eventHandler(repoPath);
         handler.handleDirectoryDeleteDiffs("directory");
         await waitFor(1);
-        // Verify file has been renamed in the shadow repo
-        expect(fs.existsSync(cacheFilePath)).toBe(true);
-        // Verify correct diff file has been generated
-        let diffFiles = fs.readdirSync(diffsRepo);
-        expect(diffFiles).toHaveLength(1);
-        const diffFilePath = path.join(diffsRepo, diffFiles[0]);
-        const diffData = readYML(diffFilePath);
-        expect(diffData.source).toEqual(DIFF_SOURCE);
-        expect(diffData.is_deleted).toBe(true);
-        expect(diffData.is_rename).toBeFalsy();
-        expect(diffData.is_new_file).toBeFalsy();
-        expect(diffData.created_at).toBeTruthy();
-        expect(diffData.repo_path).toEqual(repoPath);
-        expect(diffData.branch).toEqual(DEFAULT_BRANCH);
-        expect(diffData.file_relative_path).toEqual(relFilePath);
-        expect(diffData.diff).toEqual("");
+        expect(assertFileDeleteEvent(repoPath, relFilePath)).toBe(true);
     });
 
     test("with file already in .deleted",  async () => {